comm message integration

+++ apps/commMC_app/fsw/inc/commMC_app_extern_typedefs.h
@@ -15,7 +15,8 @@
 
 typedef struct
 {
-    uint16 OutMsgToSend; /**< \brief Command Task ID to specify the action to be taken */
+    int OutMsgToSend; /**< \brief Command Task ID to specify the action to be taken */
+    
 } COMMMC_APP_CommandPacket_Payload_t;
 
 typedef struct
